refactor: Move Header.tsx file.
Move from src/app/components/navigation to src/app/components
